Added category management in the admin panel

+++ modules/admin/controllers/CategoryController.php
@@ -2,6 +2,7 @@
 
 namespace app\modules\admin\controllers;
 
+use app\modules\admin\models\product;
 use Yii;
 use app\modules\admin\models\Category;
 use app\modules\admin\models\CategorySearch;
@@ -104,8 +105,16 @@ class CategoryController extends AppAdminController
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        $cats = Category::find()->where(['parent_id' => $id])->count();
+        $products = Product::find()->where(['category_id' => $id])->count();
+        if ($cats || $products) {
+            Yii::$app->session->setFlash('error', 'Удаление данной категории невозможно: к категории приклеплены другие категории или товары');
 
+        }else {
+            $this->findModel($id)->delete();
+            Yii::$app->session->setFlash('success', 'Категория удалена');
+
+        }
         return $this->redirect(['index']);
     }
